#96 doing some more filter profiling. hot spot is simply in the OnePoleFilter sample processor math. I have ideas.... test project builds again, fixing some build warnings

+++ src/clarinoid/basic/Stopwatch.hpp
@@ -74,7 +74,7 @@ struct TimeSpan
         return (float)d;
     }
     float ElapsedBeats(float bpm) const {
-        return double(mMicros) * bpm / 60000000;
+        return (float)(double(mMicros) * bpm / 60000000);
     }
 
     bool IsZero() const
